Sign-up Page: elements Fixes #9
