--- frontend/src/components/SignupFormPage.js
@@ -1,85 +1,0 @@
-import { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { Link, Redirect, useHistory } from "react-router-dom";
-import { signup, demoLogin } from "../store/session";
-import './css/SignupForm.css';
-
-function SignupFormPage() {
-	const dispatch = useDispatch();
-	const history = useHistory();
-	const sessionUser = useSelector((state) => state.session.user);
-	const [email, setEmail] = useState("");
-	const [username, setUsername] = useState("");
-	const [password, setPassword] = useState("");
-	const [confirmPassword, setConfirmPassword] = useState("");
-	const [errors, setErrors] = useState([]);
-
-	if (sessionUser) return <Redirect to="/app/all" />;
-
-	const handleSubmit = (e) => {
-		e.preventDefault();
-		if (password === confirmPassword) {
-			setErrors([]);
-			return dispatch(signup({ email, username, password })).then(() => {
-				history.push("/app/all");
-			}).catch(async (res) => {
-				const data = await res.json();
-				if (data && data.errors) setErrors(data.errors);
-			});
-		}
-		return setErrors(['Confirm Password field must be the same as the Password field']);
-	};
-
-	return (
-		<>
-			<h1>Sign Up</h1>
-			<form onSubmit={handleSubmit}>
-				<ul>
-					{errors.map((error, idx) => <li key={idx}>{error}</li>)}
-				</ul>
-				<label>
-					Email
-					<input
-						type="text"
-						value={email}
-						onChange={(e) => setEmail(e.target.value)}
-						required
-					/>
-				</label>
-				<label>
-					Username
-					<input
-						type="text"
-						value={username}
-						onChange={(e) => setUsername(e.target.value)}
-						required
-					/>
-				</label>
-				<label>
-					Password
-					<input
-						type="password"
-						value={password}
-						onChange={(e) => setPassword(e.target.value)}
-						required
-					/>
-				</label>
-				<label>
-					Confirm Password
-					<input
-						type="password"
-						value={confirmPassword}
-						onChange={(e) => setConfirmPassword(e.target.value)}
-						required
-					/>
-				</label>
-				<button type="submit">Sign Up</button>
-			</form>
-			<div onClick={e => dispatch(demoLogin()).then(() => history.push("/app/all"))}>Demo Login</div>
-			<span>Already have an account?</span>
-			<Link to="/login">Sign in</Link>
-		</>
-	);
-}
-
-export default SignupFormPage;
