+++ frontend/src/components/css/ListDetails.css
@@ -1,7 +1,7 @@
 #list-details-container {
 	z-index: 0;
 	left: 55.1%;
-	width: 45.1%;
+	width: 44%;
 	background-color: #e4eef8;
 	margin-left: -3px;
 	padding-left: 3px;
