+++ frontend/src/components/css/ListView.css
@@ -1,5 +1,5 @@
 #list-view-container {
-	height: 100%;
+	height: 99%;
 	width: 12%;
 	margin-right: 0.5%;
 	margin-left: 0.5%;
