+++ frontend/src/components/css/SignupPage.css
@@ -1,0 +1,11 @@
+#auth-container {
+	display: flex;
+}
+
+#left-container {
+	width: 50%;
+}
+
+#right-container {
+	width: 50%;
+}
