+++ frontend/src/components/css/TaskDetails.css
@@ -1,6 +1,6 @@
 #task-details-container {
 	background-color: white;
-	width: 45.1%;
+	width: 45%;
 	height: 99%;
 	margin-top: 0.5%;
 	border-radius: 5px;
